More movement of control messages
