+++ ros_sensors/main.py
@@ -25,6 +25,7 @@ from affect import Affect
 from audio_power import AudioPower
 from chat_track import ChatTrack
 from control import Control
+from control_psi import ControlPsi
 from face_track import FaceTrack
 from sound_track import SoundTrack
 from room_brightness import RoomBrightness
@@ -37,6 +38,7 @@ logging.info("Starting the OpenCog ROS Bridge")
 print "Starting the OpenCog ROS Bridge"
 
 co = Control()
+cp = ControlPsi()
 af = Affect()
 ap = AudioPower()
 ct = ChatTrack()
